Got compensation angle working!
I had to refactor the code some to make it more manageable, placing almost the calculation in a function outside of the main game loop.

Maybe I will make another very similar one to model the enemy behaviour.

Also I am successfully computing the angle to the following checkpoint
and have done some thrust management based on the facing angle.

Tried to implement simple slow down when approaching target but seeing the opponents AI deal with this problem tells me that this wont be nearly enough. I need to anticipate the next target and start accelerating towards that before I even make the checkpoint.
The opponent seems get his body inside the edge of the checkpoint which is much more efficient. It must have some kind of model based on its momentum and when it is sure it will pass by the checkpoint it shuts off engine and starts facing its next target. Once it has got the checkpoint it starts again.